+++ src/service/task/aggregateEventReservations.ts
@@ -1,4 +1,3 @@
-import * as cinerino from '@cinerino/domain';
 import * as factory from '@tokyotower/factory';
 
 import { IConnectionSettings, IOperation } from '../task';
@@ -6,6 +5,7 @@ import { IConnectionSettings, IOperation } from '../task';
 import { RedisRepository as EventWithAggregationRepo } from '../../repo/event';
 import { MongoRepository as PerformanceRepo } from '../../repo/performance';
 import { RedisRepository as CheckinGateRepo } from '../../repo/place/checkinGate';
+import { MongoRepository as ProjectRepo } from '../../repo/project';
 import { MongoRepository as ReservationRepo } from '../../repo/reservation';
 
 import * as AggregateService from '../aggregate';
@@ -19,7 +19,7 @@ export function call(data: factory.task.aggregateEventReservations.IData): IOper
             checkinGate: new CheckinGateRepo(settings.redisClient),
             eventWithAggregation: new EventWithAggregationRepo(settings.redisClient),
             performance: new PerformanceRepo(settings.connection),
-            project: new cinerino.repository.Project(settings.connection),
+            project: new ProjectRepo(settings.connection),
             reservation: new ReservationRepo(settings.connection)
         });
     };